Create Version Comp to add it in the About tab
